# openapi/fleetd-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "fleetd API",
    "description": "Versioned control-plane contract for cooperating software agents.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.10.0"
  },
  "paths": {
...
    "/v1/channels": {
      "get": {
        "tags": [
          "channels"
        ],
        "summary": "List channels",
        "description": "Operator-only.",
        "operationId": "listChannels",
        "responses": {
          "200": {
            "description": "Channels",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Channel"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operator credential required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "channels"
        ],
        "summary": "Create a channel",
        "description": "Operator-only. Initial membership is committed with the channel.",
        "operationId": "createChannel",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChannel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Channel created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Channel"
                }
              }
            }
          },
          "400": {
            "description": "Invalid channel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operator credential required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Initial member not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Channel conflicts with existing state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/channels/{channel_id}": {
      "patch": {
        "tags": [
          "channels"
        ],
        "summary": "Rename a shared channel",
        "description": "Operator-only. Direct conversation names are derived from their participants; archived channels are immutable.",
        "operationId": "renameChannel",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "description": "Shared channel ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameChannel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Renamed channel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Channel"
                }
              }
            }
          },
          "400": {
            "description": "Invalid channel name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operator credential required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Channel not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Direct, archived, or duplicate channel name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/channels/{channel_id}/archive": {
      "post": {
        "tags": [
          "channels"
        ],
        "summary": "Archive a shared channel",
        "description": "Operator-only and idempotent. Archive is one-way, retains permanent membership and immutable history, and rejects new messages.",
        "operationId": "archiveChannel",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "description": "Shared channel ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Archived channel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Channel"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operator credential required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Channel not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Direct conversations cannot be archived",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/channels/{channel_id}/members": {
      "get": {
        "tags": [
          "channels"
        ],
        "summary": "List exact channel memberships",
        "description": "Available to the operator or a member of this exact channel. The bounded projection omits opaque agent metadata.",
        "operationId": "listChannelMembers",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exact channel memberships",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChannelMember"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Channel membership required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Channel not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "channels"
        ],
        "summary": "Add a channel member",
        "description": "Operator-only. Membership is permanent for the channel lifetime.",
        "operationId": "addChannelMember",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMember"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Member added or already present"
          },
          "401": {
            "description": "Missing or invalid credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operator credential required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Channel or agent not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Existing membership uses another delivery mode",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/v1/channels/{channel_id}/read-cursor": {
      "put": {
        "tags": [
          "channels"
        ],
        "summary": "Advance a personal conversation read cursor",
        "description": "Agent-only and monotonic. The authenticated participant may acknowledge only a cursor already committed in this channel; stale retries cannot move the cursor backwards.",
        "operationId": "advanceConversationRead",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceConversationRead"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated personal attention state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationAttention"
                }
              }
            }
          },
          "400": {
            "description": "Negative or future read cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Agent credential and channel membership required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Channel not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/v1/conversations/attention": {
      "get": {
        "tags": [
          "channels"
        ],
        "summary": "List personal conversation attention",
        "description": "Agent-only. Returns exact unread and explicitly addressed counts derived from the authenticated participant's durable membership cursors and immutable channel messages.",
        "operationId": "listConversationAttention",
        "responses": {
          "200": {
            "description": "Personal attention state",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConversationAttention"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Agent credential required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Channel": {
        "type": "object",
        "description": "A durable conversation shared by a bounded set of agents.",
        "required": [
          "id",
          "name",
          "kind",
          "metadata",
          "created_at_ms"
        ],
        "properties": {
          "archived_at_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "created_at_ms": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ConversationKind"
          },
          "metadata": {},
          "name": {
            "type": "string"
          }
        }
      },
...
      "CreateChannel": {
        "type": "object",
        "description": "Input for creating a shared channel and its initial membership.",
        "required": [
          "name"
        ],
        "properties": {
          "member_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateChannelMember"
            }
          },
          "metadata": {},
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Stable JSON envelope returned for fleetd domain errors.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}